requirements.txt
numpy
pandas
requests
python-dotenv
sqlalchemy
psycopg2-binary
temp
apache-airflow-providers-postgres